Refactored Twitter data provider
