+++ yeltzland/Models/Tweet.swift
@@ -1,0 +1,13 @@
+//
+//  Tweet.swift
+//  yeltzland
+//
+//  Created by John Pollard on 24/07/2020.
+//  Copyright © 2020 John Pollard. All rights reserved.
+//
+
+import Foundation
+
+struct Tweet {
+    var fullTweet: String
+}
